[구아현] 코드 위치 이동
