+++ src/main/java/site/metacoding/recipemarket/web/UserController.java
@@ -3,11 +3,13 @@ package site.metacoding.recipemarket.web;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -15,6 +17,8 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import lombok.RequiredArgsConstructor;
 import site.metacoding.recipemarket.service.UserService;
+import site.metacoding.recipemarket.util.UtilValid;
+import site.metacoding.recipemarket.web.dto.user.JoinReqDto;
 import site.metacoding.recipemarket.web.dto.user.RememberReqDto;
 
 @RequiredArgsConstructor
@@ -25,6 +29,17 @@ public class UserController {
     private final UserService userService;
     private final HttpServletResponse response;
     private final HttpServletRequest request;
+
+    // 회원가입
+    @PostMapping("/join")
+    public String join(@Valid JoinReqDto joinReqDto, BindingResult bindingResult) {
+
+        UtilValid.요청에러처리(bindingResult);
+
+        userService.회원가입(joinReqDto.toEntity());
+
+        return "redirect:/login-form";
+    }
 
     // 메인 페이지
     @GetMapping("/")
