+++ src/main/resources/templates/user/joinForm.mustache
@@ -15,7 +15,7 @@
     </style>
 
 </head>
- 
+
 <body>
     <h2>회원가입 페이지</h2>
     <div class="container">
